Added check for the support of -fwrapv.
This adds the AC macro LD_CHECK_CC_SWITCH to check if the compiler supports
the given command line switch. To enable a defined wrap-around behaviour for
signed integers (currently needed), we add -fwrapv to CFLAGS if supported by
the compiler (#635).
LD_CHECK_CC_SWITCH gets a list of variables as second argument and adds the
switch to all of them, if the switch is supported.
(src/autoconf/configure.in)

Signed-off-by: zesstra <zesstra@zesstra.de>

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2602 56160b6e-f8f8-0310-bd54-bccebfe6a2f2