Use mmap() for getting memory from the system (smalloc.c)
sbrk() and brk() are obsoleted and some systems (like Darwin) don't have a
working implementation anymore. If sbrk/brk() are not usable, but mmap()
is available, use mmap() for mapping anonymous memory which we then use.
If mmap() is not available, use the old fallback (system malloc()).
Changes some precompiler checks for SBRK_OK, which should be REPLACE_MALLOC.
(smalloc.c)
Signed-off-by: zesstra <zesstra@zesstra.de>

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2605 56160b6e-f8f8-0310-bd54-bccebfe6a2f2