Removed define free_string_svalue()
The define free_string_svalue() was used to replace the function calls to
that function in interpret.c by a direct call to free_mstring(), skipping
the check for the correct type if compiled with DEBUG. I think we should
remain consistent here and check as well in interpret.c. Added INLINE as well.
(interpret.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2631 56160b6e-f8f8-0310-bd54-bccebfe6a2f2