BC cleanup: Removed some macros, used dedicated offset type instead of int32.
We just write all data as unsigned in the bytecode. Once the data was written,
the interpretation as signed/unsigned is anyway up to the code which reads it.
The _LONG macros and _long() functions from bytecode.h were  rarely used and
replaced by fixed-sized types with specific semantic.
Use dedicated offset type instead of int32.

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2671 56160b6e-f8f8-0310-bd54-bccebfe6a2f2