Identifiers now use hash16_t instead of short (#597).
The struct ident_s used a signed short as hash type, which limited the
hash table size for identifiers to 32668 hash chains.
The hash type was changed to hash16_t which does not change the ident_s
size but allows for larger tables.
Additionally the lexxer uses hash_string() from mstrings instead of
using the low-level hash functions.
(lex.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2679 56160b6e-f8f8-0310-bd54-bccebfe6a2f2