Added the function index to the function header.
To check argument types at runtime, we need the function index in the
defining program. Sometime we only have the pointer to the function
header and it is non-trivial to get the function index from this. With
the index stored in the header we can access the function information in
the defining progp.
(exec.h, prolang.y)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2780 56160b6e-f8f8-0310-bd54-bccebfe6a2f2