Optimize checks for destructed objects in mappings.
A lot of operations (prominently: sizeof(mapping)) require that mappings
do not contain references to destructed objects and call
check_map_for_destr() before doing their work. In case of sizeof() you
expect a very fast efun with complexity O(1), but in fact it has O(n)
and sizeof() can take quite some time for large mappings. Checks for
destructed objects should only be done when necessary.
This patch introduces a global counter which is increment on each
destructed object and a mapping member which records the value of this
counter at the time of the last check for destructed objects.
check_map_for_destr() immediately returns if the stamp in the mapping is
equal to the global counter.

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2861 56160b6e-f8f8-0310-bd54-bccebfe6a2f2