Changed from alarm() to setitimer(ITIMER_REAL,)
The alarm had to be (re-)set every time after an alarm occurred and if an
alarm was missed somehow, this could lead to switched off alarms (main
reason for check_alarm).
We now use setitimer(ITIMER_REAL,...) to set a recurring realtime alarm,
which simplifies matters and should make the check_alarm() obsolete.
(backend.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2873 56160b6e-f8f8-0310-bd54-bccebfe6a2f2