Fixed control stack handling in check_function_args().
check_function_args() failed to take another two special situations
into account: at the start of top-level evaluations the value of inter_pc
is one from the last evaluation and has to be set.
Additionally, no frame must be removed if there are only two and the
bottom one is a dummy frame.
(interpret.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2908 56160b6e-f8f8-0310-bd54-bccebfe6a2f2