Improve handling of SIGINT, SIGTERM.
The handlers for SIGINT and SIGTERM are installed with SA_RESETHAND, which
restores the default signal handler upon signal delivery. Then a repeated
signal is handled immediately with the default action and in the next
backend cycle. In case of SIGINT our signal handler will be re-installed
in the next backend cycle if the mudlib master handles the signal
properly.
(backend.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2922 56160b6e-f8f8-0310-bd54-bccebfe6a2f2