Synchronize the alarm timer to full seconds.
As the mud internal time is only updated each alarm time this resulted in
up to (alarm_time+1) seconds delay between the two. When the first
timer is set up, we now take care to expire it on a full second to improve
synchronicity (to a max. difference of (alarm_time)s).
(backend.c)

git-svn-id: svn://svn.bearnip.com/ldmud/3.3/trunk@2925 56160b6e-f8f8-0310-bd54-bccebfe6a2f2