Detect long executions with ITIMER_PROF.
ITIMER_PROF is a timer which accounts time spent on the process and its
syscalls. This patch adds a handler for the SIGPROF signal and sets a
timer at the beginning of each top-level execution. At the end the
timer is deactivated again.
Should it trigger in between the handler will set a flag and the
interpreter will dump a stacktrace into the debuglog after the current
instruction.
This feature is deactivated by default.
(backend.c, interpreter.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2930 56160b6e-f8f8-0310-bd54-bccebfe6a2f2