Introduced new float format using native doubles.
Up to now the driver used a 48 bit wide float consisting of a int32_t
mantissa and a int16_t exponent.
The new format FLOAT_FORMAT_2 uses native doubles in the svalue,
simplifying code and increases precision.
The disadvantage as a larger memory footprint of svalues. Therefore the
new float format is only enabled on LP64 architectures, where p_int
(and the LPC int) are 64 bit wide numbers, because on this systems the
increase in memory consumption is limited to the bytecode (+2 bytes per
float).
For storing the doubles with full precision a new savefile format was
created (version 2). It stores the doubles with snprintf using "%la"
which stores the double without loss in precision. Additionally it makes
the code simpler and the savefiles more compact.
If values are stored in format 0 and 1 on drivers which use native doubles
a loss in precision will take place.

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@3033 56160b6e-f8f8-0310-bd54-bccebfe6a2f2