json_serialize(): fixed mapping serialization.
Call walk_mapping() once per mapping, not for every key in it.
Don't serialize random values in mappings with width 0. In fact,
raise an error for mappings with width 0, because JSON does not
support keys without values.
Instead of silently ignoring any but the first value in mappings with
width > 1, raise an error for these mappings as JSON does not support it
and serializing several values in an array would ambigious.

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@3041 56160b6e-f8f8-0310-bd54-bccebfe6a2f2