Changed default for data cleanup time.
The time between two data cleanups of an objects was derived from the
general cleanup time (time until to call clean_up() in the object).
This did not make much sense. It now has a default of 10800s. It can
only be changed at compile-time, because configure_driver() does not
exist in 3.3.x.