Changed restoring integers to use strtoimax().
strtoimax() has the benefit of setting errno to ERANGE when the value
does not fit into intmax_t. If that is the case of the value does not
fit into p_int, we truncate and issue a warning.
Additionally fixed the final value of <pt> which was not set right when
restoring integer numbers.
(fixes #791)