Added tests for issue #791.
One test tries to restore a too large integer on ILP32 systems
and checks if the truncation warning is raised.
The second test tries to restore a too large integer (on all
supported platforms) and checks for the truncation warning.
BTW: an array of the last runtime warnings was made available
in base.inc. Might be used in the future for other stuff as
well.