Fixed crash during compact_mapping().
LDMud could crash in svalue_cmp() during compact_mapping(), when
the number of values in the hash part of the mapping is not
correct.
The happened, because resize_mapping (used for copy_mapping(),
copy(), m_reallocate()) skipped keys referencing destructed
objects, but did not correct the number of keys in the hash part,
resulting in a hash part having less keys than stated in .hash.used.
(fixes #816)