setup_new_frame() passes allowRefs to setup_new_frame2.
Currently references in varargs arguments are only allowed for
simul-efuns and strangely for lambda closures in driver hooks.
To implement a default function in the master (as opposed to an lfun
in the object itself) it must be allowed to call lfun closures with
references in varargs parameters (because I know of no way to create
lambdas with varargs arguments). This patch implements that by passing
the allowRefs argument from int_call_lambda via setup_new_frame
to setup_new_frame2.