Don't warn about inconsistent prototypes when weak-types are in effect.
With #pragma weak_types there is no effect when inconsistent prototypes
occur, as the argument types are not checked or saved. Also unknown
functions will prototyped just by calling them resulting in different
views on the number of arguments (for a possibly varargs function).
Therefore the warning about the deactivation of type checks should
not occur with weak_types.