Comparison failed when no type information was available for both operands.
When there is no type information about both operands check_binary_op_types()
took a shortcut and returned error_type, because there is nothing to test.
But on assignment the error_type as NULL and the assignment operation checked
to result type against NULL to throw an additional error.
So check_binary_op_types() must not return NULL for success.