Introduced a default type for functions.
Functions without a type now implicitly get the type "mixed"
(lpctype_mixed).
Function types being NULL are a nasty source of NULL pointer
dereferences because this is apparantly the instance a type
can be NULL. It was forgotten in the past and may be forgotten
in future code. To prevent that, we want to ensure, the functions
(internally) always have a type.
(Fixed #837)