Moved check for return type consistency up.
This check is now being done before the argument checks,
this gives at least in some cases more meaningful
messages.

Also, we require now that prototypes and function definition
have the exact same return type. In the case of overloading an
inherited function the type must only be compatible.

On the cosmetic side, the patch also changes some variables
to more meaningful names.