Check for inconsistent arguments types improved.
The argument checks trigger an error if the argument types
are not the same in prototype and function definition.
During redefinition of inherited functions different
argument types are allowed if they are compatible. If
not, the behaviour depends on pragmas. pedantic makes
them an error, check_overloads makes them a warning.