Use a common type object for all structs of the same name.
Instead of creating a type object for each struct definition,
create one only for its name, so it'll stay constant even on
recompilation of the program. This is done by introducing a
new type struct_name_t that contains the program and struct
name and will be hashed (instead of the struct_name_t before).