Refill lexer input buffer in time after processing newline characters.
The lexer input buffer is refilled in blocks containing whole lines.
So each newline character may also represent the end of the input
buffer (then it will be followed by a '\0' character), wherefore
there needs to be a check for it and if so the buffer be refilled.
This patch added the required checks to three places that handle '\n'.
(Fixes #833)