Fix creating a loop in a union list during GC.
The driver keeps all union types having the same head (=union type with
all the types except the last one) in a singly-linked list. GC rebuilds
that list and usually processes each type object only once. This didn't
work for <float|int> which is a static type object (because it's used
in compiler type lookup tables). The check that should have prevented
that type being but in the list twice didn't work, when it was the
first type being put in that list (thus always having .next == NULL).