Ignore some auto-generated files in git
Files generated by the build process are ignored
by git now.
However, files generated by autoconf, automake or
with the help of git etc. remain in the repository
for people not having these tools (because they
just downloaded a tagged archive and want to use
configure as is).
They will be updated before tagging releases.