Corrected the variable index adjustments in symbol_variable().
As current_variables points to the non-virtual variables now,
at was not sufficient to subtract it from the local variable index.
For non-virtual variables the index needed to be considered relative
to the first non-virtual variable in the current program.
For virtual variables the variable's origin needed the be looked
up in the current object's inherit list to make the necessary
adjustments.