Correct the type comparision between struct types
lpctype_contains() checks whether <src> can be stored in <dest> by comparing
the relationship between them. It checked whether <src> is a base of <dest>
(or equal to <dest>), but it must be the other way around (the more
specific/derived type can be stored in the more general/base type).