Better type deduction for runtime member lookups for structs
For runtime member lookups the compiler can't determine the destination
member and its type. So it used to consider the result as being mixed.
But it will be one of the members, so we can make a union of all
the member types and then have a more specific result type.
(Fixes the old test scenario 040413.)