Check argument types against definition, not prototypes.
When doing a call_other the RTT checks checked the actual
arguments against the argument types in the destination
object's program, not the inherited program that contained
the actual definition of the function. They may be different
if the inheriting program contains a different prototype
(and may lead to a crash if that prototype has less arguments
than the definition - which is allowed for varargs functions).