Make DH parameters configurable
The DH parameters can now be set with configure_driver().
The argument is expected to be a PEM encoded string or
0 to restore the built-in defaults.
pkg-gnutls keeps the parameters in a permanently allocated
block, pkg-openssl does not keep a copy of the parameters
for itself.
If the parameters can't be imported, the driver tries to
retain the old parameters, which is guaranteed for GnuTLS,
but not for OpenSSL.
(fixes #846)