prevent generation of SIGPIPE signals
If supported (very recent GnuTLS), use GNUTLS_NOSIGNAL
(MSG_NOSIGNAL) to not generate SIGPIPE when writing.
If supported (MacOS), set SO_NOSIGPIPE on stream sockets
to not generate SIGPIPE when writing.
Actually, we ignore SIGPIPE anyway, but there were sporadic
problems with this reported on MacOS. Also, the debugger
still stops on the signal by default, even if ignored.