telnet_neg(): removed compiler warnings.
The variable for holding a char is now an unsigned char.
It was anyway limited to 0xff and the sprintf() did
handle it as a char, not an int.
Sine NTELOPTS ist 256 by default, we have to disable a
compiler warning that a comparison of unsigned char < 256
is always true.