Allow more complex lvalues.
Until now lvalue index/range expressions could only consist
of a base-lvalue (that is not an index/range) and the index
expression. With this commit the indexed lvalue expression
can be equally complex. To do this it is not only sufficient
to store one instruction to convert an rvalue to an lvalue,
but a duplicate of the entire expression bytecode as an lvalue.