Fixed memory leak in allocate() in out of memory situations.
allocate() with non-trivial init values leaks the copies of that
init value that were already made in the event of an out-of-memory error.
Because freeing of complex structures requires memory itself (for the
serialization of these structures), instead of freeing the (partly)
array at once added a cleanup handler that reverses the allocate()
algorithm freeing each element seperately.