Prevent execution of replace_program on destructed objects
When doing a replace_program lambda closures are recompiled
and differences flagged with a runtime error or warning.
But now lambda() uses the rvalues and thus gets zeros
for destructed objects, so any closures or other object
references within the lambda structure will turn up as zero
und thus will lead to a different lambda and therefore an error.