Prevent input handling for destructed players
When a player is destructed don't to any telnet, input_to or command handling
for it, because it may result in calling a function into a destructed
object (and thus yield a fatal error). Usually destructed players are
removed before the call of get_message(), but newly created connections
or handling telnert negotations may result in freshly destructed objects
in the list of open connections.