Deactivate stack mismatch check on a return from a lambda closure
In a lambda closure a return opcode can occur in the middle of an expression,
where evaluated parts of the expression are already put on the stack,
resulting in a fatal error indicating that the stack head is higher than
the function frame. Although the message is correct it is normal behaviour
for lambda closures, therefore this check is now deactivated there.
The superfluous values on the stack will be removed later anyway.
(Fixes #868)