Do not replace the system malloc by default.
We tried to replace the system malloc & Co by our own
allocator if we deemed this safe. However, our test
did not seem to detect that reliably. Also, our
memory allocate does not meet the requirements of a
general memory allocator (e.g. alignment).
Therefore we do not try to replace malloc by default
anymore.
However, if someone wants to try it, there is a
configuration switch for it.
(Fixes #862)