Fixed Define __MAX_MALLOC__ to be dynamic
Since the hard memory limit is dynamically adjustable,
the define for it must also be dynamic.
Also changed function names to include suffix _buf to
stress that they return a freshly allocated buffer,
not a simple number.
And finally used a constant buffer size (21) and snprintf
to guarantee the buffer does not overflow.