Fixed a bug that would delay handling of charmode input.
The telnet machine indicates commands ready to process with the TS_READY
state. This does not work for charmode, because even when some characters
were received the telnet machine might be just within a telnet sequence
and therefore needs the state variable to indicate that. Earlier driver
versions saved the telnet state elsewhere in this case and the variable
was set to TS_READY nevertheless. This mechanism was removed when
streamlining and documenting the telnet machine for 3.6, therefore the
get_message() function was not aware of any ready character and handled
only one of them per backend cycle.

Corrected this by introducing a new telnet state TS_CHAR_READY and
preventing the telnet machine to process telnet sequences before the
chars have been processed.