Corrected size calculations in terminal_colour()
terminal_colour() works in 2 phases, in the first the size of the resulting
string is calculated, in the second the string is generated. A bugfix for
the wrapping algorithm to reset the position of the last space after a
newline was only done for the second phase, resulting in a different
size calculation in the first phase. (Fixes #872)