Nicer error message of a declarative cast to the same type
Declarative casts must always be from a more general to a specialized type.
When the type in the cast is the same as the type of the value, this rule
is also violated and the error message speaks of general and specialized
types. If the types are indeed the same we can just give a message about
being the same type.