Corrected offsets when patching simul-efun call_other calls
For simul-efun calls the F_SAVE_ARG_FRAME opcode is inserted after
the object and function name of the call_other have already been parsed.
If the simul-efun is called via call_direct() also the simul-efun file
and function name have to be insert before that. For this the location
of the opcodes for the first argument (the object name) has to be used,
but wasn't.