Fix handling of star operator in regular expressions
The star operator is implemented by taking as many characters as possible
and trying to match the rest, backtracking by one character if the match
fails. As the matching code keeps state in global variables, the attempt
of matching the remaining expression changes the global state and was not
properly reset upon backtracking. (Fixes #878)