Don't split off too small blocks in slaballoc/smalloc
When the allocator finds a block larger than needed, it splits off
the remaining block. If the remaining block is too small to be allocated
as a large block, it will be marked as allocated, but not
garbage-collectable and left there for all eternity. Also the corresponding
magic bytes were not set, resulting in a crash when the garbage collector
checks them.

Now, if the block is too small, abstain from splitting and return the
whole (larger) block, so that the whole space can be reclaimed later on.