baseof() and to_struct() accept base/sub structs with the same name
Structs with a different definition but the same name (for example when
a program might have been recompiled) are usually considered the same.
baseof() and to_struct() will now apply this rule also when searching for
the base struct. If to_struct() encounters such different definitions it
will convert to the required struct by copying each member by name.
(Fixes #884)