Update other struct's base struct when reactivating a struct
When a struct is reactivated, update any definitions that use this struct
as a base. Since 3.5.0 the members don't need to be updated, because the
definition is not part of the type anymore. But the update of the base
was also erroneously thrown out at that time, thus preventing the
reactivation of those derived structs. (Fixes #884)