Fix lambda() not building argument frame for #'call_direct
An argument frame (F_SAVE_ARG_FRAME) is needed for all efuns with a variable
number of arguments. For tabled efuns this is detected by make_func, but
call_other() and call_direct() are not tabled, but also have a variable
number of arguments. lambda() took special care of call_other(), but not
of call_direct().