Fix assertion failure in certain virtual inherit scenarios
When a program - that has already an old inherit - inherits another
program that has an updated inherit (i.e. the old one and a new one),
the compiler tried to update the second old one again, resulting in an
assertion failure (because it didn't expect that is has already an update).