Implement prefix #'lfun:: and #'var:: for closure literals
Closure literals can now use a prefix lfun:: to create lfun closures
and var:: to create variable closures. Because variable closures have
the lowest priority, the override is required when there are (s)efuns
or lfuns with the name. An override for lfuns is not needed (they have
the highest priority), but is added for completeness' sake.