sprintf() added a suppressed newline from column mode later on
sprintf() suppresses the final newline from a string printed in column mode,
when that table (all columns) is followed by a newline in the format string.
(So there will be only one newline and not both.) But a status variable
was not cleared, so the newline came back later on and was printed at the
next newline or format string end. (Fixes #529)