Prevent telnet_neg() from advancing beyond the current input
telnet_neg() is not only responsible for telnet negotiations but also
deals with line endings. For this to work properly, it checks whether
the interactive is in character mode. Since the mode may be switched
by the current command, it is important not to process input further
beyond that point.

To this end, we split telnet_neg() into two parts, telnet_neg_reset()
that is responsible for switching back to TS_DATA mode, and telnet_neg()
that does the actual processing. This allows resetting the telnet
neg state without further processing. It also makes telnet_neg()
almost idempotent: if called in TS_READY or TS_CHAR_READY state, it
will have no effect.

We must be careful to advance the telnet machinery on each iteration
through the interactives in get_message(); otherwise, processing
of commands for players will be slowed down to 1 per second.
The t-comm-receive-bulk test checks that this doesn't happen.

Fixes #892.