Fix function name refcounting during compilation
During compilation the function name string is not refcounted for inherited
functions (because those strings will not be part of the final program).
When a function switches its status the refcount must be adjusted
accordingly. This can happen for virtually inherited functions that vanish,
because undefined functions get a short function definition (with just
F_UNDEFINED) at the end of the compilation and thus will be part of the
final program.