Merge the simul_efunp and simul_efun_table tables
The simul_efunp was designed to be used for simul-efuns that have an
index > 65535, but nothing in LDMud supports those large indices.
Therefore there is no need for a separate table. Instead we're using
a magic index to indicate that the simul-efun is not in the table and
should be called by name.