simulate: add parse_error line to master log_error
When `src/simulate.c`'s `parse_error` function is invoked to handle
a compiler error or warning the line number the event occurred at is
provided to the `parse_error` func as the `int line` arg.

Prior to this commit the line number available in `parse_error` was not
provided to the `log_error` function of the in-game master ob. In this
commit we add this as an argument to the `log_error` apply so lib code
can provide better error messages, or file context, as part of the
`log_error` response.