Fix movement of variable blocks during virtual inheritance
When a virtual inherit is updated to a newer inherit, its variable block
usually changes in size and therefore subsequent blocks need to be moved.
This was not done for duplicate entries in the inherit list, but needs
to be done for them as well, because later routines depend on those
inherit structures having correct variable information.
(Fixes #899)