bug: account for negative start offset in efun::read_bytes
Prior to this commit when `efun::read_bytes` is provided a negative
start offset that would extend beyond the beginning of the file, 0 was
returned.

It's more desirable to instead start at offset `0` and read as much of
the provided range as possible. This facilitates (for e.g.)
`sefun::tail` to provide a fixed negative start offset to
`efun::read_bytes` while still functioning for files smaller than that
fixed negative offset in total size.