Introduce recursion limit in traditional regexp engine
The traditional regular expression engine does a recursive call upon each
decision it must make when matching a text. So the recursion depth is
dependent on the size of this text and therefore may easily exceed the
available stack space. Therefore introducing a build-time configuration for
a maximum recursion limit that will cause an LPC runtime error when hit.