main: mention JSON support in --options output.
Prior to this commit even when `ldmud` was built with JSON support
enabled, it wasn't reported in `ldmud --options` along with other
optional features like XML or IPv6 support.

This commit updates the main.c implementation of the `--options` output
to include a line indicating JSON support when `USE_JSON` is defined.