python: handle Python 3.10 func type annotations.
In Python 3.10 the `PyFunction_GetAnnotations` function used  by
`python_register_efun` to find the Python types associated with an efun
being registered changed. Previously it returned a dict that allowed
lookup by name. In Python 3.10's C API it is returning a tuple and has
to be accessed by numeric index.

This commit introduces a small helper function `python_find_type_in_seq`
that can be used to iterate the result from `PyFunction_GetAnnotations`
as a generic sequence, trying to find a type by name. This is in turn
used by `python_register_efun` when it needs to find a function's return
types, or the types for its arguments.

Applying this commit fixes the `t-python` unit test when linking Python
3.10+.