tests: add startup.py ldmud.Object invocation.
It's possible to create a reference to an LPC object from Python using
`ldmud.Object("...")` during `startup.py` initialization. Crucially this
happens before the master object is fully initialized. With LDMud 3.6.4
this early object reference works as expected. With LDMud 3.6.5 it
provokes a segfault.

This commit introduces a unit test that can provoke the segfault with
3.6.5 and ensure a subsequent fix doesn't regress.