Fixed false warning in unique_array()
unique_array() issued a warning that the function (when given as a string)
doesn't exist in the current object. This was due to the fact, that a
callback object was created temporarily with the current object before
using it on every object given in the first argument. As we don't care
about whether this function exists on every object, disable this warning
for unique_array(). (Fixes #903)