Call always __INIT for lighweight objects
When creating, copying or restoring lightweight objects always call __INIT.
This is needed, because restore will not touch nosave variables, and __INIT
might have side-effects that are needed.

Also when copying and restoring free any previous contents of variables
(might come from __INIT or the UID hook). And when pragma share_variables
is in effect, abort __INIT also in lightweight objects.