Fixed cleanup of coroutines on error
When handling a runtime error only the current coroutine was aborted (and
corresponding awaiting coroutines), but not any other coroutines in the
call stack (i.e. coroutines that used call_coroutine() to start an
independent coroutine), with the effect that running coroutines were freed
(resulting in an assertion failure and awaiting coroutines being locked in
that state).