Unify global identifier handling in the compiler
The LPC compiler needs to differentiate between global identifiers that are
for the current program only (lfuns, global variables, structs), and
identifiers that are driver-global (efuns, simul-efuns). The former need to
be removed after compilation, the latter only cleaned.

The handling of this was spread throughout the code and done in different
ways, and sometimes not done at all (eg. for struct definitions).