lp-245: fix misc warnings in preloaded code.
This commit fixes a handful of warnings that occur when running the
`mud/lp-245` sample MUD lib with the current driver version. I only went
as far as to fix warnings in code that is preloaded at startup. More
warnings may exist in parts of the lib that are not listed in
`mud/lp-245/room/init_file`.