Do object-wise cleanups
When doing a garbage collection the driver does a mud-wide cleanup first.
For this cleanup a pointer table to note each encountered value is created.
This table uses a massive amount of memory which usually results in a crash
due to out-of-memory condition when a garbage collection is really needed.
Therefore we now reset this table for each object. This will result in
duplicate cleanup effort for shared values, but uses less memory.