Fix crash in the GC when printing a closure to an LWO
When the GC encountered a closure to an LWO it tries to print the
program name of the LWO. But the program might have been collected
in the same GC run and therefore already be overwritten with memory
management structures. The GC would then follow illegal pointers.
Fixed also similar cases in the GC.