doc: Add coroutine info to driver_info manpage.
In 1f36a5203 coroutine statistics were implemented, including adding
constants (`DI_NUM_COROUTINES`, & `DI_SIZE_COROUTINES`) to use with
`efun::driver_info` to fetch a count of the number of coroutines loaded,
and their size in memory. This commit adds a mention of these two
`what` values in the man page for the `driver_info` efun.