Add coroutine info to driver_info() manpage
Previously coroutine statistics were implemented, including adding
constants (DI_NUM_COROUTINES and DI_SIZE_COROUTINES) to use with
efun::driver_info() to fetch a count of the number of coroutines loaded,
and their size in memory. This commit adds a mention of these two
values in the man page for the driver_info() efun.