autoconf/configure.ac: Update to autoconf 2.71
The original autoconf script was probably produced with autoconf version
1.8, and various changes have been made to it over the years without
running autoconf -u (update).  This is problematic because autoconf
macros are not guaranteed to be stable between releases.  Consequently,
a number of oddities have crept in over time.  This commit fixes that by
ensuring that the configure.ac file is compliant with the current release
of autoconf.

Regenerate pre-prepared configure script as well, for those who do not
have autoconf.

Additionally, a few places were using uint and u_long, which are not part of the
C99 standard.  Replace them with, respectively, unsigned int and unsigned long.