Fix compiler warnings
In ed.c fix 'storing the address of local variable' by making the indentation
stack static. In bitstrings.c and comm.c fix 'variable set but not used'.
In md5.c GCC's warning of 'accessing 64 bytes in a region of size 15' seems
to be a false positivie, therefore disabling the warning there. In xalloc.c
disable GCC's warning of 'storing the address of local variable' as this
is intended for stack safety checks.