Fix sign issue when reading cross-definition offsets from function flags.
C23 changed the type of enumeration constants to be the same as the enum
type (instead of always being 'int'). This meant that now INHERIT_MASK is
regarded as being unsigned int, instead of (signed) int. As a consequence
functions that calculated the cross-definition offsets were losing the sign,
resulting in wrong offsets and potentially crashing.